+++ lib/src/internal/main/stats/stat_manager.dart
@@ -492,6 +492,7 @@ class StatManager {
     int? logiTs,
     int? errorCode,
     String? errorDescription,
+    String? connectionId,
   }) async {
     sbLog.d(StackTrace.current);
 
@@ -515,7 +516,7 @@ class StatManager {
         'logi_latency': logiLatency,
         'error_code': errorCode,
         'error_description': errorDescription,
-        'connection_id': const Uuid().v1(),
+        'connection_id': connectionId ?? const Uuid().v1(),
       },
     );
 
@@ -550,7 +551,7 @@ class StatManager {
   }) {
     sbLog.d(StackTrace.current);
 
-    if (_wsConnectStartTsMap[endpoint] == null) {
+    if (_apiResultStartTsMap[endpoint] == null) {
       _apiResultStartTsMap[endpoint] = DateTime.now().millisecondsSinceEpoch;
     }
   }