+++ lib/src/internal/network/http/http_client/request/user/count/user_unread_message_count_request.dart
@@ -7,6 +7,7 @@ import 'package:sendbird_chat_sdk/src/internal/network/http/http_client/http_cli
 import 'package:sendbird_chat_sdk/src/internal/network/http/http_client/request/api_request.dart';
 import 'package:sendbird_chat_sdk/src/public/main/define/enums.dart';
 import 'package:sendbird_chat_sdk/src/public/main/model/message/unread_message_count.dart';
+import 'package:sendbird_chat_sdk/src/public/main/params/channel/feed_channel_total_unread_message_count_params.dart';
 import 'package:sendbird_chat_sdk/src/public/main/params/channel/group_channel_total_unread_message_count_params.dart';
 
 class UserTotalUnreadMessageCountGetRequest extends ApiRequest {
@@ -15,19 +16,32 @@ class UserTotalUnreadMessageCountGetRequest extends ApiRequest {
 
   UserTotalUnreadMessageCountGetRequest(
     Chat chat, {
-    GroupChannelTotalUnreadMessageCountParams? params,
+    GroupChannelTotalUnreadMessageCountParams? groupChannelParams,
+    FeedChannelTotalUnreadMessageCountParams? feedChannelParams,
     String? userId,
   }) : super(chat: chat, userId: userId) {
     url = 'users/${getUrlEncodedUserId(chat, userId)}/unread_message_count';
-    final customTypes = params?.channelCustomTypes ?? [];
-    final superFilter = params?.superChannelFilter ?? SuperChannelFilter.all;
+
+    final customTypes = groupChannelParams?.channelCustomTypes ?? [];
+    final superFilter =
+        groupChannelParams?.superChannelFilter ?? SuperChannelFilter.all;
+    final isFeedChannel =
+        chat.chatContext.appInfo?.notificationInfo?.feedChannels.isNotEmpty ??
+            false;
+
     queryParams = {
       if (customTypes.isNotEmpty) 'custom_types': customTypes,
       'super_mode': groupChannelSuperFilterEnum(superFilter),
-      if (chat.chatContext.appInfo?.notificationInfo?.feedChannels.isNotEmpty ??
-          false)
-        'include_feed_channel': true,
     };
+
+    if (isFeedChannel) {
+      queryParams['include_feed_channel'] = true;
+
+      final feedChannelKeysFilter = feedChannelParams?.feedChannelKeysFilter;
+      if (feedChannelKeysFilter != null && feedChannelKeysFilter.isNotEmpty) {
+        queryParams['feed_channel_keys'] = feedChannelKeysFilter;
+      }
+    }
   }
 
   @override