+++ lib/src/public/main/params/channel/feed_channel_total_unread_message_count_params.dart
@@ -1,0 +1,9 @@
+// Copyright (c) 2025 Sendbird, Inc. All rights reserved.
+
+/// Represents a feed channel total unread message count params.
+/// @since 4.5.0
+class FeedChannelTotalUnreadMessageCountParams {
+  /// The feed channel keys filter.
+  /// @since 4.5.0
+  List<String>? feedChannelKeysFilter;
+}